+++ demisto_sdk/__main__.py
@@ -931,7 +931,15 @@ def secrets(config, **kwargs):
     default="from-yml",
     help="The docker image to check package on. Can be a comma separated list of Possible values: 'native:maintenance', 'native:ga', 'native:dev',"
     " 'all', a specific docker image from Docker Hub (e.g devdemisto/python3:3.10.9.12345) or the default"
-    " 'from-yml'.",
+    " 'from-yml', 'native:target'. To run lint only on native supported content with a specific image,"
+    " use 'native:target' with --docker-image-target <specific-image>.",
+)
+@click.option(
+    "-dit",
+    "--docker-image-target",
+    default="",
+    help="The docker image to lint native supported content with. Should only be used with "
+    "--docker-image native:target. An error will be raised otherwise.",
 )
 @click.option(
     "-cdam",
@@ -993,6 +1001,7 @@ def lint(**kwargs):
         coverage_report=kwargs.get("coverage_report"),  # type: ignore[arg-type]
         docker_timeout=kwargs.get("docker_timeout"),  # type: ignore[arg-type]
         docker_image_flag=kwargs.get("docker_image"),  # type: ignore[arg-type]
+        docker_image_target=kwargs.get("docker_image_target"),  # type: ignore[arg-type]
         time_measurements_dir=kwargs.get("time_measurements_dir"),  # type: ignore[arg-type]
     )