+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fixed an issue where running **validate** with the `-g` flag would skip some validations for old format files.
 * Deprecated integrations and scripts will not run anymore when providing the **--all-packs** to the **lint** command.
 * Fixed an issue where a pack `serverMinVersion` is calculated by its content items to be the minimum fromVersion.
+* Added the `--docker-image-target` flag to **lint** for testing native supported content with new images.
 
 ## 1.10.5
 * Fixed an issue where running **run-test-playbook** would not use the `verify` parameter correctly. @ajoga